+++ tests/test_dbp.py
@@ -2,33 +2,23 @@ from networkx import DiGraph
 
 import dbp
 
-MAKEFILE = dbp.MAKE_HEAD
-MAKEFILE += """\t1 \\
-\t2
-1/${STAMP}: 2/${STAMP}
-"""
-MAKEFILE += dbp.MAKE_TAIL
+MAKEFILE = """.PHONY: all
+all:
+ALL_REPOS = \\
+	1 \\
+	2
+.1.stamp: .2.stamp
+ALL_REPO_STAMPS := $(patsubst %,.%.stamp,${ALL_REPOS})
+TIMESTAMP = $(shell date '+%F %T')
 
+all: ${ALL_REPO_STAMPS}
 
-def test_deb_build_options_string():
-    assert dbp.deb_build_options_string(False) == "-e=DEB_BUILD_OPTIONS="
-    assert (
-        dbp.deb_build_options_string(True) == "-e=DEB_BUILD_OPTIONS=nostrip noopt debug"
-    )
-    assert dbp.deb_build_options_string(False, 5) == "-e=DEB_BUILD_OPTIONS=parallel=5"
-    assert (
-        dbp.deb_build_options_string(True, 5)
-        == "-e=DEB_BUILD_OPTIONS=nostrip noopt debug parallel=5"
-    )
-
-
-def test_irun():
-    assert dbp.irun(["true"], True) == 0
-    assert dbp.irun(["true"], False) == 0
-    assert dbp.irun(["false"], True) == 1
-    assert dbp.irun(["false"], False) == 1
-    assert dbp.irun([], True) == -1
-    assert dbp.irun([], False) == -1
+${ALL_REPO_STAMPS}: REPO = $(@:.%.stamp=%)
+${ALL_REPO_STAMPS}: LOG = ${REPO}.log
+${ALL_REPO_STAMPS}:
+\t@echo ${TIMESTAMP} Starting dbp build ${REPO}
+\t@dbp --cname "$${USER}-dbp-parallel-${REPO}" build ${REPO} >${LOG} 2>&1
+\t@: >$@"""
 
 
 def test_makefile():
@@ -36,4 +26,4 @@ def test_makefile():
     g.add_node(1)
     g.add_node(2)
     g.add_edge(1, 2)
-    assert dbp.makefile(g) == MAKEFILE
+    assert dbp.generate_makefile(g) == MAKEFILE